removed experimental tool MarkDuplicatesGATK in favor of the accurate Picard implemenation and MarkDuplicatesGATK
