+++ src/test/java/org/broadinstitute/hellbender/tools/walkers/markduplicates/AbstractMarkDuplicatesCommandLineProgramTest.java
@@ -1,17 +1,23 @@
 package org.broadinstitute.hellbender.tools.walkers.markduplicates;
 
 import htsjdk.samtools.*;
+import htsjdk.samtools.metrics.MetricsFile;
 import htsjdk.samtools.util.CloseableIterator;
 import htsjdk.samtools.util.CloserUtil;
 import org.apache.commons.io.FileUtils;
+import org.apache.spark.SparkException;
 import org.broadinstitute.hellbender.CommandLineProgramTest;
 import org.broadinstitute.hellbender.cmdline.CommandLineProgram;
 import org.broadinstitute.hellbender.cmdline.StandardArgumentDefinitions;
+import org.broadinstitute.hellbender.engine.ReadsDataSource;
+import org.broadinstitute.hellbender.exceptions.UserException;
 import org.broadinstitute.hellbender.testutils.ArgumentsBuilder;
 import org.broadinstitute.hellbender.testutils.BaseTest;
 import org.broadinstitute.hellbender.testutils.IntegrationTestSpec;
 import org.broadinstitute.hellbender.testutils.SamAssertionUtils;
-import org.broadinstitute.hellbender.testutils.testers.AbstractMarkDuplicatesTester;
+import org.broadinstitute.hellbender.testutils.testers.MarkDuplicatesSparkTester;
+import org.broadinstitute.hellbender.utils.read.GATKRead;
+import org.broadinstitute.hellbender.utils.read.markduplicates.GATKDuplicationMetrics;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
@@ -19,9 +25,14 @@ import picard.sam.util.PhysicalLocationInt;
 import picard.sam.util.ReadNameParser;
 
 import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
  * This class defines the individual test cases to run.  The actual running of the test is done
@@ -31,7 +42,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     public static final File TEST_DATA_DIR = new File(getTestDataDir(), "walkers/MarkDuplicatesGATK/");
 
-    protected abstract AbstractMarkDuplicatesTester getTester();
+    protected abstract MarkDuplicatesSparkTester getTester();
 
     protected abstract CommandLineProgram getCommandLineProgramInstance();
 
@@ -45,14 +56,14 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testSingleUnmappedFragment() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addUnmappedFragment(-1, DEFAULT_BASE_QUALITY);
         tester.runTest();
     }
 
     @Test
     public void testTwoUnmappedFragments() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addUnmappedFragment(-1, DEFAULT_BASE_QUALITY);
         tester.addUnmappedFragment(-1, DEFAULT_BASE_QUALITY);
         tester.runTest();
@@ -60,21 +71,21 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testSingleUnmappedPair() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addUnmappedPair(-1, DEFAULT_BASE_QUALITY);
         tester.runTest();
     }
 
     @Test
     public void testSingleMappedFragment() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedFragment(1, 1, false, DEFAULT_BASE_QUALITY);
         tester.runTest();
     }
 
     @Test
     public void testTwoMappedFragments() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedFragment(0, 1, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedFragment(0, 1, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.runTest();
@@ -82,14 +93,14 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testSingleMappedPair() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(1, 1, 100, false, false, DEFAULT_BASE_QUALITY);
         tester.runTest();
     }
 
     @Test
     public void testSingleMappedFragmentAndSingleMappedPair() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedFragment(1, 1, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addMappedPair(1, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.runTest();
@@ -97,7 +108,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairs() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(1, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedPair(1, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.runTest();
@@ -105,7 +116,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testThreeMappedPairs() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(1, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedPair(1, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addMappedPair(1, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
@@ -114,7 +125,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testSingleMappedFragmentAndTwoMappedPairs() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedFragment(1, 1, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addMappedPair(1, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedPair(1, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
@@ -123,7 +134,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsAndTerminalUnmappedFragment() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(1, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedPair(1, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addUnmappedFragment(-1, DEFAULT_BASE_QUALITY); // unmapped fragment at end of file
@@ -132,7 +143,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsAndTerminalUnmappedPair() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(1, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedPair(1, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addUnmappedPair(-1, DEFAULT_BASE_QUALITY); // unmapped pair at end of file
@@ -141,7 +152,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testOpticalDuplicateFinding() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
 
         // explicitly creating 1 expected optical duplicate pair
         tester.setExpectedOpticalDuplicate(1);
@@ -158,7 +169,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testOpticalDuplicatesDifferentReadGroups() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.setExpectedOpticalDuplicate(0);
         tester.addMatePair("RUNID:7:1203:2886:82292",  19, 19, 485253, 485253, false, false, true, true, "42M59S", "59S42M", true, false, false, false, false, DEFAULT_BASE_QUALITY, "H0164.2");  // duplicate
         tester.addMatePair("RUNID:7:1203:2886:82242", 19, 19, 485253, 485253, false, false, false, false, "42M59S", "59S42M", true, false, false, false, false, DEFAULT_BASE_QUALITY, "H0164.1");
@@ -174,7 +185,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testOpticalDuplicatesTheSameReadGroup() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.setExpectedOpticalDuplicate(1);
         tester.addMatePair("RUNID:7:1203:2886:82292",  19, 19, 485253, 485253, false, false, true, true, "42M59S", "59S42M", true, false, false, false, false, DEFAULT_BASE_QUALITY, "H0164.2");  // duplicate (tie broken by readname)
         tester.addMatePair("RUNID:7:1203:2886:82242", 19, 19, 485253, 485253, false, false, false, false, "42M59S", "59S42M", true, false, false, false, false, DEFAULT_BASE_QUALITY, "H0164.2");
@@ -191,7 +202,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testOpticalDuplicatesAndPCRDuplicatesOrientationDifference() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.setExpectedOpticalDuplicate(0);
         tester.addMatePair("RUNID:7:1203:2886:16838",  19, 19, 485253, 486253, false, false, true, true, "101M", "101M", true, false, false, false, false, DEFAULT_BASE_QUALITY, "1");  // duplicate
         tester.addMatePair("RUNID:7:1203:2886:16834", 19, 19, 486253, 485253, false, false, false, false, "101M", "101M", false, true, false, false, false, DEFAULT_BASE_QUALITY, "1");
@@ -204,7 +215,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
     // to FR if they point in opposite directions. This test asserts that two pairs of reads that ultimately all map to 93470499-93470499 with clipping
     // and orientations will ultimately be marked and mapped together even though they have opposite orientations (one is FR while the other is RF).
     public void testReadSameStartPositionOrientationOverride() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.setExpectedOpticalDuplicate(0);
         tester.addMatePair("HJYFJCCXX160204:8:1124:31659:21526",  19, 19, 93470380, 93470499, false, false, false, false, "31S111M9S", "64M87S", true, false, false, false, false, DEFAULT_BASE_QUALITY, "1");  // after clipping these both start on 93470499
         tester.addMatePair("HJYFJCCXX160204:8:1124:31659:21527",  19, 19, 93470499, 93470380, false, false, true, true, "64M87S", "31S111M9S", false, true, false, false, false, DEFAULT_BASE_QUALITY, "1");  // after clipping these both start on 93470499
@@ -215,7 +226,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
     // This asserts that we are flipping reads in the Pair object based on both start position and contig index, this does not
     // make a difference unless the start position is the same across two contigs, so we assert it is handled properly
     public void testReadsHaveSameStartPositionButDifferentChromosomeNonEquivalence() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.setExpectedOpticalDuplicate(0);
         tester.addMatePair("HJYFJCCXX160204:8:1124:31659:21526",  19, 20, 93470380, 93470499, false, false, false, false, "31S111M9S", "64M87S", true, false, false, false, false, DEFAULT_BASE_QUALITY, "1");  // after clipping these both start on 93470499
         tester.addMatePair("HJYFJCCXX160204:8:1124:31659:21527",  20, 19, 93470499, 93470380, false, false, true, true, "64M87S", "31S111M9S", false, true, false, false, false, DEFAULT_BASE_QUALITY, "1");  // after clipping these both start on 93470499
@@ -224,7 +235,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testOpticalDuplicateClusterSamePositionNoOpticalDuplicatesWithinPixelDistance() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(101);
         tester.setExpectedOpticalDuplicate(0);
         tester.addMatePair("RUNID:7:1203:2886:16834", 1, 485253, 485253, false, false, true, true, "42M59S", "59S42M", false, true, false, false, false, DEFAULT_BASE_QUALITY); // duplicate
@@ -234,7 +245,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testOpticalDuplicateClusterSamePositionOneOpticalDuplicatesWithinPixelDistance() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(45);
         tester.setExpectedOpticalDuplicate(1);
         tester.addMatePair("RUNID:7:1203:2886:16834", 1, 485253, 485253, false, false, true, true, "45M", "45M", false, true, false, false, false, DEFAULT_BASE_QUALITY); // duplicate
@@ -244,7 +255,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testOpticalDuplicateClusterOneEndSamePositionOneCluster() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(101);
         tester.setExpectedOpticalDuplicate(1);
         tester.addMatePair("RUNID:7:2205:17939:39728", 1, 485328, 485312, false, false, false, false, "55M46S", "30S71M", false, true, false, false, false, ELIGIBLE_BASE_QUALITY);
@@ -254,7 +265,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsAndMappedSecondaryFragment() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(1, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedPair(1, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addMappedFragment(1, 200, false, DEFAULT_BASE_QUALITY, true); // mapped non-primary fragment
@@ -263,7 +274,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMappedFragmentAndMappedPairFirstOfPairNonPrimary() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMappedFragment(1, 1, false, ELIGIBLE_BASE_QUALITY);
         tester.addMatePair(1, 200, 0, false, true, false, false, "54M22S", null, false, false, true, true, false, DEFAULT_BASE_QUALITY);
@@ -272,7 +283,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsMatesSoftClipped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMappedPair(1, 10022, 10051, false, false, "76M", "8S68M", false, true, false, DEFAULT_BASE_QUALITY);
         tester.addMappedPair(1, 10022, 10063, false, false, "76M", "5S71M", false, true, false, DEFAULT_BASE_QUALITY);
@@ -282,7 +293,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsWithSoftClipping() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         //todo Picard says no duplicates, we say duplicate, what's going on?
         // NB: no duplicates
@@ -295,7 +306,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsWithSoftClippingFirstOfPairOnlyNoMateCigar() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.setNoMateCigars(true);
         // NB: no duplicates
@@ -308,7 +319,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsWithSoftClippingBoth() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         // mapped reference length: 73 + 42 = 115
         tester.addMappedPair(1, 10046, 10002, true, true, "3S73M", "6S42M28S", true, false, false, DEFAULT_BASE_QUALITY); // duplicate
@@ -319,7 +330,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMatePairSecondUnmapped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMatePair(1, 10049, 10049, false, true, false, false, "11M2I63M", null, false, false, false, false, false, DEFAULT_BASE_QUALITY);   // neither are duplicates
         tester.runTest();
@@ -327,7 +338,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMatePairFirstUnmapped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMatePair(1, 10056, 10056, true, false, false, false, null, "54M22S", false, false, false, false, false, DEFAULT_BASE_QUALITY);    // neither are duplicates
         tester.runTest();
@@ -335,7 +346,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMappedFragmentAndMatePairSecondUnmapped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMatePair(1, 10049, 10049, false, true, false, false, "11M2I63M", null, false, false, false, false, false, ELIGIBLE_BASE_QUALITY);
         // We set the length here separately because some of the MD variants use TOTAL_MAPPED_REFERENCE_LENGTH as their DUPLICATE_SCORING_STRATEGY.  If we kept the read length
@@ -347,7 +358,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMappedFragmentAndMatePairFirstUnmapped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMatePair(1, 10049, 10049, true, false, false, false, null, "11M2I63M", false, false, false, false, false, ELIGIBLE_BASE_QUALITY);
            // We set the length here separately because some of the MD variants use TOTAL_MAPPED_REFERENCE_LENGTH as their DUPLICATE_SCORING_STRATEGY.  If we kept the read length
@@ -359,7 +370,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMappedPairAndMatePairSecondUnmapped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMatePair(1, 10040, 10040, false, true, true, false, "76M", null, false, false, false, false, false, DEFAULT_BASE_QUALITY); // second a duplicate,
         // second end unmapped
@@ -369,7 +380,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMappedPairAndMatePairFirstUnmapped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMatePair(1, 10040, 10040, true, false, false, true,  null, "76M", false, false, false, false, false, DEFAULT_BASE_QUALITY); // first a duplicate,
         // first end unmapped
@@ -380,7 +391,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
     // TODO: fails on MarkDuplicatesWithMateCigar
     @Test
     public void testMappedPairAndMatePairFirstOppositeStrandSecondUnmapped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(101);
         // first end mapped OK -, second end unmapped
         tester.addMatePair(1, 484071, 484071, false, true, false, false,  "66S35M", null, true, false, false, false, false, DEFAULT_BASE_QUALITY);
@@ -389,9 +400,126 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
         tester.runTest();
     }
 
+    @Test( dataProvider = "md")
+    public void testMarkDuplicatesSparkIntegrationTestLocal(
+        final File input, final long totalExpected, final long dupsExpected,
+        Map<String, List<String>> metricsExpected) throws IOException {
+
+        ArgumentsBuilder args = new ArgumentsBuilder();
+        args.add("--"+ StandardArgumentDefinitions.INPUT_LONG_NAME);
+        args.add(input.getPath());
+        args.add("--"+StandardArgumentDefinitions.OUTPUT_LONG_NAME);
+
+        File outputFile = createTempFile("markdups", ".bam");
+        outputFile.delete();
+        args.add(outputFile.getAbsolutePath());
+
+        args.add("--"+StandardArgumentDefinitions.METRICS_FILE_LONG_NAME);
+        File metricsFile = createTempFile("markdups_metrics", ".txt");
+        args.add(metricsFile.getAbsolutePath());
+
+        runCommandLine(args.getArgsArray());
+
+        Assert.assertTrue(outputFile.exists(), "Can't find expected MarkDuplicates output file at " + outputFile.getAbsolutePath());
+
+        int totalReads = 0;
+        int duplicateReads = 0;
+        try ( final ReadsDataSource outputReads = new ReadsDataSource(outputFile.toPath()) ) {
+            for ( GATKRead read : outputReads ) {
+                ++totalReads;
+
+                if ( read.isDuplicate() ) {
+                    ++duplicateReads;
+                }
+            }
+        }
+
+        Assert.assertEquals(totalReads, totalExpected, "Wrong number of reads in output BAM");
+        Assert.assertEquals(duplicateReads, dupsExpected, "Wrong number of duplicate reads in output BAM");
+
+        final MetricsFile<GATKDuplicationMetrics, Comparable<?>> metricsOutput = new MetricsFile<>();
+        try {
+            metricsOutput.read(new FileReader(metricsFile));
+        } catch (final FileNotFoundException ex) {
+            System.err.println("Metrics file not found: " + ex);
+        }
+        final List<GATKDuplicationMetrics> nonEmptyMetrics = metricsOutput.getMetrics().stream().filter(
+                metric ->
+                    metric.UNPAIRED_READS_EXAMINED != 0L ||
+                    metric.READ_PAIRS_EXAMINED != 0L ||
+                    metric.UNMAPPED_READS != 0L ||
+                    metric.UNPAIRED_READ_DUPLICATES != 0L ||
+                    metric.READ_PAIR_DUPLICATES != 0L ||
+                    metric.READ_PAIR_OPTICAL_DUPLICATES != 0L ||
+                    (metric.PERCENT_DUPLICATION != null && metric.PERCENT_DUPLICATION != 0.0 && !Double.isNaN(metric.PERCENT_DUPLICATION)) ||
+                    (metric.ESTIMATED_LIBRARY_SIZE != null && metric.ESTIMATED_LIBRARY_SIZE != 0L)
+        ).collect(Collectors.toList());
+
+        Assert.assertEquals(nonEmptyMetrics.size(), metricsExpected.size(),
+                            "Wrong number of metrics with non-zero fields.");
+        for (int i = 0; i < nonEmptyMetrics.size(); i++ ){
+            final GATKDuplicationMetrics observedMetrics = nonEmptyMetrics.get(i);
+            List<?> expectedList = metricsExpected.get(observedMetrics.LIBRARY);
+            Assert.assertNotNull(expectedList, "Unexpected library found: " + observedMetrics.LIBRARY);
+            Assert.assertEquals(observedMetrics.UNPAIRED_READS_EXAMINED, expectedList.get(0));
+            Assert.assertEquals(observedMetrics.READ_PAIRS_EXAMINED, expectedList.get(1));
+            Assert.assertEquals(observedMetrics.UNMAPPED_READS, expectedList.get(2));
+            Assert.assertEquals(observedMetrics.UNPAIRED_READ_DUPLICATES, expectedList.get(3));
+            Assert.assertEquals(observedMetrics.READ_PAIR_DUPLICATES, expectedList.get(4));
+            Assert.assertEquals(observedMetrics.READ_PAIR_OPTICAL_DUPLICATES, expectedList.get(5));
+            Assert.assertEquals(observedMetrics.PERCENT_DUPLICATION, expectedList.get(6));
+
+            //Note: IntelliJ does not like it when a parameter for a test is null (can't print it and skips the test)
+            //so we work around it by passing in an 'expected 0L' and only comparing to it if the actual value is non-null
+            if (observedMetrics.ESTIMATED_LIBRARY_SIZE != null && (Long)expectedList.get(7) != 0L)  {
+                Assert.assertEquals(observedMetrics.ESTIMATED_LIBRARY_SIZE, expectedList.get(7));
+            }
+        }
+    }
+
+    @Test
+    public void testSupplementaryReadUnmappedMate() {
+        File output = createTempFile("supplementaryReadUnmappedMate", "bam");
+        final ArgumentsBuilder args = new ArgumentsBuilder();
+        args.addOutput(output);
+        args.addInput(getTestFile("supplementaryReadUnmappedmate.bam"));
+        runCommandLine(args);
+
+        try ( final ReadsDataSource outputReadsSource = new ReadsDataSource(output.toPath()) ) {
+            final List<GATKRead> actualReads = new ArrayList<>();
+            for ( final GATKRead read : outputReadsSource ) {
+                Assert.assertFalse(read.isDuplicate());
+                actualReads.add(read);
+            }
+
+            Assert.assertEquals(actualReads.size(), 3, "Wrong number of reads output");
+        }
+    }
+
+    @Test
+    public void testHashCollisionHandling() {
+        // This test asserts that the handling of two read pairs with the same start positions but on different in such a way
+        // that they might cause hash collisions are handled properly.
+        final File output = createTempFile("supplementaryReadUnmappedMate", "bam");
+        final ArgumentsBuilder args = new ArgumentsBuilder();
+        args.addOutput(output);
+        args.addInput(getTestFile("hashCollisionedReads.bam"));
+        runCommandLine(args);
+
+        try ( final ReadsDataSource outputReadsSource = new ReadsDataSource(output.toPath()) ) {
+            final List<GATKRead> actualReads = new ArrayList<>();
+            for ( final GATKRead read : outputReadsSource ) {
+                Assert.assertFalse(read.isDuplicate());
+                actualReads.add(read);
+            }
+
+            Assert.assertEquals(actualReads.size(), 4, "Wrong number of reads output");
+        }
+    }
+
     @Test
     public void testMappedPairAndMappedFragmentAndMatePairSecondUnmapped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMatePair(1, 10040, 10040, false, true, true, false, "76M", null, false, false, false, false, false, DEFAULT_BASE_QUALITY); // first a duplicate,
         // second end unmapped
@@ -402,7 +530,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMappedPairAndMappedFragmentAndMatePairFirstUnmapped() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMatePair(1, 10040, 10040, true, false, false, true, null, "76M", false, false, false, false, false, DEFAULT_BASE_QUALITY); // first a duplicate,
         // first end unmapped
@@ -413,7 +541,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsWithOppositeOrientations() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMappedPair(1, 10182, 10038, true, true, "32S44M", "66M10S", true, false, false, DEFAULT_BASE_QUALITY); // -/+
         tester.addMappedPair(1, 10038, 10182, false, false, "70M6S", "32S44M", false, true, false, ELIGIBLE_BASE_QUALITY); // +/-, both are duplicates
@@ -422,7 +550,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsWithOppositeOrientationsNumberTwo() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMappedPair(1, 10038, 10182, false, false, "70M6S", "32S44M", false, true, false, ELIGIBLE_BASE_QUALITY); // +/-, both are duplicates
         tester.addMappedPair(1, 10182, 10038, true, true, "32S44M", "66M10S", true, false, false, DEFAULT_BASE_QUALITY); // -/+
@@ -431,7 +559,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testThreeMappedPairsWithMatchingSecondMate() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         // Read0 and Read2 are duplicates
         // 10181+41=10220, 10058
@@ -445,7 +573,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMappedPairWithSamePosition() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMappedPair(1, 4914, 4914, false, false, "37M39S", "73M3S", false, false, false, DEFAULT_BASE_QUALITY); // +/+
         tester.runTest();
@@ -453,7 +581,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMappedPairWithSamePositionSameCigar() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMappedPair(1, 4914, 4914, false, false, "37M39S", "37M39S", false, false, false, DEFAULT_BASE_QUALITY); // +/+
         tester.runTest();
@@ -461,7 +589,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairWithSamePosition() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMappedPair(0, 5604914, 5604914, false, false, "37M39S", "73M3S", false, false, false, ELIGIBLE_BASE_QUALITY); // +/+
         tester.addMappedPair(0, 5604914, 5604914, true, true, "37M39S", "73M3S", false, false, false, DEFAULT_BASE_QUALITY); // +/+
@@ -470,7 +598,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairWithSamePositionDifferentStrands() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(0, 5604914, 5604914, false, false, "50M", "50M", true, false, false, ELIGIBLE_BASE_QUALITY); // +/-
         tester.addMappedPair(0, 5604914, 5604914, true, true, "50M", "50M", false, true, false, DEFAULT_BASE_QUALITY); // -/+
         tester.runTest();
@@ -478,7 +606,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairWithSamePositionDifferentStrands2() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(0, 5604914, 5604915, false, false, "50M", "50M", true, false, false, ELIGIBLE_BASE_QUALITY); // +/-
         tester.addMappedPair(0, 5604915, 5604914, true, true, "50M", "50M", false, true, false, DEFAULT_BASE_QUALITY); // -/+
         tester.runTest();
@@ -486,7 +614,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testMappedPairWithFirstEndSamePositionAndOther() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(76);
         tester.addMappedPair(0, 5604914, 5605914, false, false, "37M39S", "73M3S", false, false, false, DEFAULT_BASE_QUALITY); // +/+
         tester.addMappedPair(0, 5604914, 5604914, false, false, "37M39S", "73M3S", false, false, false, DEFAULT_BASE_QUALITY); // +/+
@@ -495,7 +623,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoGroupsOnDifferentChromosomesOfTwoFragments() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedFragment(0, 1, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedFragment(0, 1, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addMappedFragment(1, 1, false, ELIGIBLE_BASE_QUALITY);
@@ -506,7 +634,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoGroupsOnDifferentChromosomesOfTwoMappedPairs() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(0, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedPair(0, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addMappedPair(1, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
@@ -516,7 +644,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoGroupsOnDifferentChromosomesOfThreeMappedPairs() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(0, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedPair(0, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addMappedPair(0, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
@@ -528,7 +656,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testThreeGroupsOnDifferentChromosomesOfThreeMappedPairs() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.addMappedPair(0, 1, 100, false, false, ELIGIBLE_BASE_QUALITY);
         tester.addMappedPair(0, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
         tester.addMappedPair(0, 1, 100, true, true, DEFAULT_BASE_QUALITY); // duplicate!!!
@@ -543,7 +671,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testBulkFragmentsNoDuplicates() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(100);
         for(int position = 1; position <= 10000; position += 1) {
             tester.addMappedFragment(0, position, false, "100M", DEFAULT_BASE_QUALITY);
@@ -553,7 +681,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testBulkFragmentsWithDuplicates() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(100);
         for(int position = 1; position <= 10000; position += 1) {
             tester.addMappedFragment(0, position, false, "100M", ELIGIBLE_BASE_QUALITY);
@@ -567,7 +695,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testStackOverFlowPairSetSwap() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(68);
 
         final File input = new File(TEST_DATA_DIR, "markDuplicatesWithMateCigar.pairSet.swap.sam");
@@ -583,7 +711,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testSecondEndIsBeforeFirstInCoordinate() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(68);
         tester.addMappedPair(0, 108855339, 108855323, false, false, "33S35M", "17S51M", true, true, false, DEFAULT_BASE_QUALITY); // +/-
         tester.runTest();
@@ -591,7 +719,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testPathologicalOrderingAtTheSamePosition() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(68);
 
         tester.setExpectedOpticalDuplicate(1);
@@ -620,7 +748,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testDifferentChromosomesInOppositeOrder() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(101);
         tester.setExpectedOpticalDuplicate(1);
         tester.addMatePair("RUNID:6:101:17642:6835", 0, 1, 123989, 18281, false, false, true, true, "37S64M", "52M49S", false, false, false, false, false, DEFAULT_BASE_QUALITY, "1");
@@ -630,7 +758,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testOpticalDuplicateClustersAddingSecondEndFirstSameCoordinate() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(68);
         tester.setExpectedOpticalDuplicate(1);
         tester.addMatePair("RUNID:1:1:15993:13361", 2, 41212324, 41212310, false, false, false, false, "33S35M", "19S49M", true, true, false, false, false, ELIGIBLE_BASE_QUALITY);
@@ -652,7 +780,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test(dataProvider = "secondarySupplementaryData")
     public void testTwoMappedPairsWithSupplementaryReads(final Boolean additionalFragSecondary, final Boolean additionalFragSupplementary, final Boolean fragLikeFirst) {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         tester.getSamRecordSetBuilder().setReadLength(68);
         tester.setExpectedOpticalDuplicate(1);
         tester.addMatePair("RUNID:1:1:15993:13361", 2, 41212324, 41212310, false, false, false, false, "33S35M", "19S49M", true, true, false, false, false, DEFAULT_BASE_QUALITY);
@@ -720,7 +848,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
 
     @Test
     public void testTwoMappedPairsWithSoftClippingFirstOfPairOnly() {
-        final AbstractMarkDuplicatesTester tester = getTester();
+        final MarkDuplicatesSparkTester tester = getTester();
         // NB: no duplicates
         // 5'1: 2, 5'2:46+73M=118
         // 5'1: 2, 5'2:51+68M=118
@@ -751,4 +879,125 @@ public abstract class AbstractMarkDuplicatesCommandLineProgramTest extends Comma
     }
 
 
+    @Test
+    public void testNonExistantReadGroupInRead() {
+        final MarkDuplicatesSparkTester tester = new MarkDuplicatesSparkTester(true);
+        tester.addMatePair("RUNID:7:1203:2886:82292",  19, 19, 485253, 485253, false, false, true, true, "42M59S", "59S42M", true, false, false, false, false, DEFAULT_BASE_QUALITY, "NotADuplicateGroup");
+        try {
+            tester.runTest();
+            Assert.fail("Should have thrown an exception");
+        } catch (Exception e){
+           Assert.assertTrue(e instanceof SparkException);
+           Assert.assertTrue(e.getCause() instanceof UserException.HeaderMissingReadGroup);
+        }
+    }
+
+    @Test
+    public void testNoReadGroupInRead() {
+        final MarkDuplicatesSparkTester tester = new MarkDuplicatesSparkTester(true);
+        tester.addMatePair("RUNID:7:1203:2886:82292",  19, 19, 485253, 485253, false, false, true, true, "42M59S", "59S42M", true, false, false, false, false, DEFAULT_BASE_QUALITY, null);
+
+        try {
+            tester.runTest();
+            Assert.fail("Should have thrown an exception");
+        } catch (Exception e){
+            Assert.assertTrue(e instanceof SparkException);
+            Assert.assertTrue(e.getCause() instanceof UserException.ReadMissingReadGroup);
+        }
+    }
+
+    @Test(dataProvider = "readNameData")
+    public void testOpticalDuplicatesTiebrokenByPhysicalLocationNotStartPosition(final String readName1, final String readName2) {
+        // This tests the readname based tiebreaking code in mark duplicates. Since it's ambiguous which read should be marked
+        // as duplicate or not if scores match we break ties by evaluating the readname for consistencies sake.
+
+        final ReadNameParser parser = new ReadNameParser();
+
+        final PhysicalLocationInt position1 = new PhysicalLocationInt();
+        final PhysicalLocationInt position2 = new PhysicalLocationInt();
+
+        parser.addLocationInformation(readName1, position1);
+        parser.addLocationInformation(readName2, position2);
+
+        final MarkDuplicatesSparkTester tester = getTester();
+        tester.getSamRecordSetBuilder().setReadLength(101);
+        tester.setExpectedOpticalDuplicate(0);
+
+        int compare = position1.tile - position2.tile;
+        if (compare == 0) {
+            compare = (short)position1.x - (short)position2.x;
+        }
+
+        if (compare == 0) {
+            compare = (short)position1.y - (short)position2.y;
+        }
+
+        final boolean isDuplicate = compare < 0;
+
+        // NOTE these reads are offset slightly but should have the same unclipped start postitions
+        tester.addMatePair(readName1, 1,2, 46,  false, false, !isDuplicate, !isDuplicate, "6S42M28S", "3S68M",  true, false, false, false, false, DEFAULT_BASE_QUALITY);
+        tester.addMatePair(readName2, 1,2, 51, false, false, isDuplicate, isDuplicate, "6S42M28S", "8S68M", true, false, false, false, false, DEFAULT_BASE_QUALITY);
+        tester.runTest();
+    }
+
+    @DataProvider
+    public Object[][] readNameData(){
+        return new Object[][]{
+                {"RUNID:7:1203:2886:82292", "RUNID:7:1205:3886:16834"},
+
+                {"RUNID:7:1203:2886:16756", "RUNID:7:1205:3886:16756"},
+                {"RUNID:7:1204:2886:16756", "RUNID:7:1205:3886:16756"},
+                {"RUNID:7:1205:2886:16756", "RUNID:7:1205:3886:16756"},
+                {"RUNID:7:1206:2886:16756", "RUNID:7:1205:3886:16756"},
+                {"RUNID:7:1207:2886:16756", "RUNID:7:1205:3886:16756"},
+
+                {"RUNID:7:1203:2886:16756", "RUNID:7:1203:4886:26756"},
+                {"RUNID:7:1203:3886:16756", "RUNID:7:1203:4886:26756"},
+                {"RUNID:7:1203:4886:16756", "RUNID:7:1203:4886:26756"},
+                {"RUNID:7:1203:5886:16756", "RUNID:7:1203:4886:26756"},
+                {"RUNID:7:1203:6886:16756", "RUNID:7:1203:4886:26756"},
+
+                {"RUNID:7:1203:2886:34756", "RUNID:7:1203:2886:36756"},
+                {"RUNID:7:1203:2886:35756", "RUNID:7:1203:2886:36756"},
+                {"RUNID:7:1203:2886:37756", "RUNID:7:1203:2886:36756"},
+                {"RUNID:7:1203:2886:38756", "RUNID:7:1203:2886:36756"},
+
+                //Added a test for tiebreaking accounting for the short casting done in picard
+                {"HK3T5CCXX160204:3:1112:11586:37067", "HK3T5CCXX160204:3:1112:11586:32144"}
+        };
+    }
+
+    @Test(dataProvider = "readNameData")
+    public void testOpticalDuplicateClusterSamePositionNoOpticalDuplicates(final String readName1, final String readName2) {
+        // This tests the readname based tiebreaking code in mark duplicates. Since it's ambiguous which read should be marked
+        // as duplicate or not if scores match we break ties by evaluating the readname for consistencies sake.
+
+        final ReadNameParser parser = new ReadNameParser();
+
+        final PhysicalLocationInt position1 = new PhysicalLocationInt();
+        final PhysicalLocationInt position2 = new PhysicalLocationInt();
+
+        parser.addLocationInformation(readName1, position1);
+        parser.addLocationInformation(readName2, position2);
+
+        final MarkDuplicatesSparkTester tester = getTester();
+        tester.getSamRecordSetBuilder().setReadLength(101);
+        tester.setExpectedOpticalDuplicate(0);
+
+        int compare = position1.tile - position2.tile;
+        if (compare == 0) {
+            compare = (short)position1.x - (short)position2.x;
+        }
+
+        if (compare == 0) {
+            compare = (short)position1.y - (short)position2.y;
+        }
+
+        final boolean isDuplicate = compare < 0;
+
+        tester.addMatePair(readName1, 1,485253, 485253, false, false, !isDuplicate, !isDuplicate, "42M59S", "59S42M", false, true, false, false, false, DEFAULT_BASE_QUALITY);
+        tester.addMatePair(readName2, 1,485253, 485253, false, false, isDuplicate, isDuplicate, "59S42M", "42M59S", true, false, false, false, false, DEFAULT_BASE_QUALITY);
+
+        tester.runTest();
+    }
 }
